monitor/requirements.txt
amadeus==4.4.0
paho-mqtt==1.5.1
